Fixed scanner memleak;added ability to run scripts
Memory leak fixed in tokenizer and added ability to run .ari scripts
(it's not picky right now in terms of script extensions but will fix
that later).

Weird bug with hash table where the word "so" isn't being
properly stored when coming first in function declaration. This is next
on the fix list.