Reworked program counter, recursion now working...
The way the instruction pointer was handled in the program had to be
completely reworked to allow for recursion. Now frames hold the program
counter and advances on the counter are done at the frame level. This
allows for the same instruction set to be called and run through with
different frames.

Tested with fibonacci sequence (test_scripts/fibo.ari), recursion is
working but some weird return side-effects I have to work through.
Additionally adding recursion ability has created a memory leak related to
ad-hoc frames for recursion functions. Will patch that with next commit.